feat: usePopupVisibleフックを追加し、PopupCardコンポーネントの表示状態を管理
